perf: server-prefetch the leave page
leave/page.tsx was a bare client-component shell, so the leave list only
fetched after the bundle downloaded and hydrated. Extract the query into
fetchLeaves(supabase), prefetch it on the server with the same queryKey,
and hand a dehydrated cache to the client via HydrationBoundary. The page
now hydrates with real rows straight from the HTML instead of a
post-hydration client fetch round-trip.

First proven instance of the eval report's #4; other list pages follow.

Closes #207
